HTML Page Built
The HTML page with the form displays when the server is started.  Add a
POST method to the form.  Need to research more on how redirect works
once the user clicks the submit button.

Function responds when submit button is clicked, displays type of HTTP
method performed but not the actual data.  Address this for future
commit.