res.writeHead is not a function
In my switch statement I was passing in too many arguments, so I think
this is what was throwing an error when the server started.

Another issue could be the ES6 arrow function.  I changed it back to ES2015
syntax and the server started and rendered the form on the client side
without an error.

I need to test this further to isolate the bug for future projects.