Too many arguments in switch statement
The issue was in my switch statement I passed in (path, req, res) and
all that was needed was a request and a response.  This is due to the
location of the path being the default '/' when the server is started.

I also switched the displayWelcomePage function back to ES6 for
consistency.  I added a basic HTTP Error 500 page if for some reason the
server can't find index.html file.